IBM DataPower and API Connect work
Two years after I suggested we start an API service delivery platform for disaster work, using Drupal, underneath the aegis of the Open Crisis organization, IBM acquired API Connect, which is doing just that in the commercial realm. API Connect is also Open Sourced to a degree, so that idea may still come to pass. Open Crisis did not live up to it’s promise (YET) and I pray it does.